media-fonts/libertinus: new package
From the project page:
The Libertinus fonts project is a fork of the Linux Libertine and Linux
Biolinum fonts. It started as an OpenType math companion of the
Libertine font family, and has grown to a full fork to address some of
the bugs in the fonts.

Package-Manager: Portage-2.3.99, Repoman-2.3.22