media-gfx/autotrace: new package
Utility for tracing bitmap images to convert them to vector images with
support for centerline tracing. Only a live ebuild is provided because
the last release (version 0.31.1) is quite old (from 2002) and requires
a number of patches to build on a modern system and to address the
numerous CVEs assigned to it.

Package-Manager: Portage-2.3.99, Repoman-2.3.22