media-gfx/pstoedit: new package
From its project page:
pstoedit converts PostScript(TM) and PDF files to other vector graphic
formats so that they can be edited graphically.

It is an optional dependency of autotrace, and the version that's
currently in the main gentoo ebuild repository (version 3.70-r1) does
not work properly.

Package-Manager: Portage-2.3.99, Repoman-2.3.22