app-editors/emacs: add native-comp support
Also nown as gccemacs, this uses native compilation of Emacs Lisp using
GCC's JIT infrastructure to increase performance. It is an experimental
feature present only on the feature/native-comp branch of the current
development version (28-vcs at the moment). The entire git repo is
fetched to be able to switch between the master and feature/native-comp
branch depending on the USE flag.

Package-Manager: Portage-2.3.103, Repoman-2.3.23