app-editors/emacs: add pgtk support
This uses GTK directly without relying in X libs, making running under
Wayland possible. It is an experimental feature, so some branch
switching is necessary as it was for native-comp until recently. When
enabling pgtk without enabling cairo I encountered compilation errors,
so the cairo USE flag is currently required for pgtk. In general, USE
flag compatibilities and dependencies necessary for pgtk are not well
tested.

Package-Manager: Portage-3.0.18, Repoman-3.0.2